[uads-024] - Moved application and interface folders into adapters
Also, deleted a duplicate of repository folder in adapters.
