+++ src/adapters/application/route/event/{id}.ts
